fix(clippy): T31.10 - resuelve los avisos de clippy del workspace del host
`cargo clippy --workspace --all-targets -- -D warnings` abortaba la
compilacion en el workspace del host (antosd, antos-protocol, builder):
83 avisos reales tras corregir los 14 iniciales de protocolo/builder que
bloqueaban incluso ver los de antosd. La mayoria son limpieza mecanica,
sin cambio de comportamiento; los que si tocan comportamiento se listan
aparte.

Mecanicos (sin cambio de comportamiento), por categoria:
- `"status" | _ => {...}` -> `_ => {...}` (14 sitios, wildcard_in_or_patterns):
  el literal era inalcanzable, el propio `_` ya lo cubria.
- `(a + b - 1) / b` -> `a.div_ceil(b)` (5 sitios, manual_div_ceil).
- `if A { if B { ... } }` -> `if A && B { ... }` /
  `"x" => { if C {...} }` -> `"x" if C => {...}` (9 sitios,
  collapsible_if/collapsible_match).
- `&["a", "b"]` -> `["a", "b"]` en `.args(...)` (4 sitios,
  needless_borrows_for_generic_args) y 2 `&workspace` -> `workspace`
  (needless_borrow).
- `s[N..].trim()` tras `starts_with(prefijo)` -> `strip_prefix` (2 sitios,
  manual_strip).
- `chunks_exact(N)` -> `as_chunks::<N>().0` (3 sitios, incluido el SHA-256
  de pkg.rs y el SHA-1 de web.rs).
- `c == 'a' || c == 'b'` -> `[a, b].contains` (4 sitios en vfs.rs,
  manual_pattern_char_comparison) y 2 `x >= A && x <= B` ->
  `(A..=B).contains(&x)` en vision/mod.rs.
- `&bytes[..] != LITERAL` -> `bytes[..] != LITERAL` (op_ref, wasm/parser.rs).
- `&[d.clone()]` -> `std::slice::from_ref(&d)` (3 sitios en vfs.rs).
- `format!("literal sin argumentos")` -> `.to_string()` / literal inline
  (3 sitios).
- `Iterator::last()` sobre split -> `.next_back()` (evita recorrer todo,
  flatpak.rs).
- `sort_by(|a,b| a.x.cmp(&b.x))` -> `sort_by_key` (2 sitios).
- `if let Some(_) = x` -> `x.is_some()`; `match { Some(p) => ..., None =>
  return None }` en bucle -> `x = x.parent()?` (2 sitios, ctx.rs y git.rs,
  identicos); `if let ... else { return None }` -> `?` (planner/ollama.rs).
- `cfg.x = v;` tras `T::default()` -> `T { x: v, ..Default::default() }`
  (2 sitios).
- `Vec` con push repetido de un runtime `count` -> `extend(repeat_n(...))`
  (wasm/vm.rs); `vec![...]` de tamano fijo -> array literal (exec/git.rs).
- `for i in 0..N { ... w[i] ... }` -> `for (i, wi) in w.iter().enumerate()`
  en el SHA-1 de web.rs (`i` seguia haciendo falta para el `match` de
  rondas, asi que el indice no desaparece del todo, pero deja de indexar
  `w` a mano).
- `#[allow(deprecated)]` en el modulo de tests de flow.rs: esos tests
  ejercitan deliberadamente la API deprecada, no es un olvido de migracion.

Con cambio de comportamiento deliberado, documentado en el propio commit:
- `Architecture::from_str` -> `Architecture::from_name` (builder/src/lib.rs):
  clippy::should_implement_trait — el nombre inducia a pensar que
  implementaba `std::str::FromStr` (que devuelve `Result`, no `Option`).
  Actualizados los dos llamadores (builder/src/main.rs y el propio test).
- `std::io::Error::new(ErrorKind::Other, ...)` -> `std::io::Error::other(...)`
  (4 sitios en builder/src/lib.rs): equivalente, mas idiomatico.
- `PackageAppType`: `impl Default` manual -> `#[derive(Default)]` +
  `#[default]` en `Cli` (protocolo/src/system.rs): mismo valor por defecto.
- `CrdtDocument::to_string()` (metodo inherente) -> `impl Display for
  CrdtDocument` (collab.rs): clippy::inherent_to_string. Ningun llamador
  cambia — `doc.to_string()` sigue funcionando via el `impl<T: Display>
  ToString for T` de la biblioteca estandar.
- `DarwinRuntime`: anadido `impl Default` que delega en `new()`
  (clippy::new_without_default); `new()` se conserva.
- `LauncherAppItem::new` (protocolo/src/launcher.rs, 9 argumentos):
  `#[allow(clippy::too_many_arguments)]` con la razon documentada en el
  propio codigo en vez de agrupar los parametros en una estructura —
  ese refactor tocaria los siete sitios de llamada en `system/barra`, un
  *workspace* aparte que este entorno no puede compilar (requiere GTK4,
  ausente aqui; ver la nota de T31.10 sobre `system/barra` en el propio
  ticket) y por tanto no podria verificar.

Archivos modificados: 34 (builder: 3, protocolo: 3, antosd: 28).

Verificacion:
- cargo clippy --workspace --all-targets -- -D warnings -> limpio (exit 0).
- cargo fmt --all -- --check -> limpio (reaplicado tras los cambios).
- cargo check --workspace -> limpio.
- cargo test --workspace -> 383/383, estable en dos ejecuciones
  consecutivas (un fallo de git::tests::test_cache_performance_under_30ms
  —sensible a tiempo, presupuesto de 30ms— en una corrida bajo carga
  paralela, reproducido en verde de forma aislada; mismo patron de
  flakiness preexistente ya documentado en T31.2, no relacionado con
  estos cambios).
- cargo clippy -p antosd --all-targets -- -D clippy::unwrap_used
  -D clippy::expect_used (puerta de T31.7) -> sigue en verde.

+++ system/protocolo/src/ui.rs
@@ -319,7 +319,7 @@ mod tests {
 
         // HUD must be centered horizontally
         assert_eq!(layout.intent_hud.width, 640);
-        assert_eq!(layout.intent_hud.x, ((1024 - 640) / 2) as i32);
+        assert_eq!(layout.intent_hud.x, (1024 - 640) / 2);
         assert_eq!(layout.intent_hud.y, 56);
 
         // Terminal window must fit within screen bounds
